==> Fluxo 4/04_Missing Analysis of the Complete Data Set.ipynb <==
## This notebook analyses missingness of the complete data set, containing all indexes

import pandas as pd
from datetime import datetime, timedelta

### Complete data set with daily granularity

complete_index =  pd.read_csv('CompleteIndexes.csv') 

at_least_one_value_missing = complete_index.apply(lambda x: (len(x)-x.count())>=1, axis=1)

at_least_one_value_missing.value_counts()

complete_index.info()

#### All samples have missing values as do all features

### Complete data set with weekly granularity

complete_weekly_index =  pd.read_csv('CompleteWeeklyIndexes.csv') 

at_least_one_value_missing = complete_weekly_index.apply(lambda x: (len(x)-x.count())>=1, axis=1)

at_least_one_value_missing.value_counts()

complete_weekly_index.info()

#### All samples have missing values as do all features

complete_weekly_index.describe()

complete_weekly_index.index()
