==> Fluxo 4/05.02_Imputation using Moving Average.ipynb <==
## This notebook inputs the missing values using Moving Average

import pandas as pd
from datetime import datetime
import numpy as np

#### Choose wether the imputation should be done on the data set with daily or weekly granularity

#INPUT, OUTPUT = 'CompleteIndexes.csv','MovingAverageImputed.csv'
INPUT, OUTPUT = 'CompleteWeeklyIndexes', 'MovingAverageWeeklyImputed'

df = pd.read_csv(INPUT+'.csv', parse_dates=True, index_col=0)
df_test = df.loc['2015-01-02':] #Selecting test data
df = df.loc['2000-01-01':'2015-01-01'] #Selecting training data
df.head()

def imput_moving_average_to_col(df_to_input, num_values):
    for index, row in df_to_input.iterrows():
        if pd.isna(row[0]):
            #print("_______________", row)
            count_prev = 0
            mean = 0
            index_prev = df_to_input.index.get_loc(index)
            while count_prev < num_values:
                index_prev -= 1
                if index_prev < 0:
                    break
                if not pd.isna(df_to_input.iloc[index_prev][0]):
                    count_prev += 1
                    #print(df_to_input.iloc[index_prev])
                    mean += df_to_input.iloc[index_prev]
            count_next = 0
            index_next = df_to_input.index.get_loc(index)
            while count_next < num_values:
                index_next += 1
                if index_next >= len(df_to_input):
                    break
                if not pd.isna(df_to_input.iloc[index_next][0]):
                    count_next += 1
                    #print(df_to_input.iloc[index_next])
                    mean += df_to_input.iloc[index_next]
            mean = mean / (count_prev+count_next)
            #print("MEAN" , mean, "----------------")
            df_to_input.at[index, df_to_input.columns[0]]  = mean
    return df_to_input

for col in df.columns:
    df_imp = imput_moving_average_to_col(df[[col]],3)
    df[col] = df_imp
    print("Column", col, "was imputed.")
df.head()

df.to_csv(OUTPUT+'_training.csv', index = True)
df_final = df

### Imputing Test Data

df = df_final
df.tail()

def imput_moving_average_to_col_test(df_to_input, num_values):
    if(pd.isna(df_to_input.iloc[-1,0])):
        mean = np.sum(df_to_input.iloc[(-num_values-1):-2,0]) / (num_values)
        df_to_input.at[df_to_input.index[-1], df_to_input.columns[0]] = mean
    return df_to_input

df_test.head()

for index, row in df_test.iterrows():
    indexes = [index]
    new_df = pd.DataFrame(data=row, index=indexes, columns=df_test.columns)
    df = pd.concat([df,new_df])

    for col in df.columns:
        df_imp = imput_moving_average_to_col_test(df[[col]],3)
        df[col] = df_imp

df = df.loc['2015-01-02':]
df.to_csv(OUTPUT+ '_test.csv', index = True)
